backend_python/requirements.txt
fastapi
uvicorn[standard]
uvicorn
pydantic
genkit
google-genai
python-dotenv
google-generativeai
python-docx
PyPDF2
python-multipart

aiofiles
supabase
flask-cors
sqlmodel
psycopg2-binary


edge-tts
aiofiles
